+++ src/homematicip/base/enums.py
@@ -185,6 +185,7 @@ class EcoDuration(AutoNameEnum):
 class SecurityZoneActivationMode(AutoNameEnum):
     ACTIVATION_WITH_DEVICE_IGNORELIST = auto()
     ACTIVATION_IF_ALL_IN_VALID_STATE = auto()
+    ACTIVATION_REQUEST_BASED = auto()
 
 
 class ClientType(AutoNameEnum):